+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * The timeline now support drag/drop operations
     * Status bar now shows the mouse position
     * When in keyframe recorning mode, the canvas outline turns red and a message appears in the status bar
+    * Status bar now shows fill/stroke colors
 * I/O:
     * Lottie output is slightly more compact
     * Added minimal support for CSS on SVG import